refactor(console): 拆分内容治理页面 helper
- 抽出内容治理选项、类型与目标展示 helper

- 保持审核、禁言、封禁和日志动作语义不变

- 降低 ModerationPage 主文件复杂度
