+++ Frontend/radish.console/src/pages/Moderation/ModerationPage.tsx
@@ -21,516 +21,49 @@ import {
   type ContentReportQueueItemVo,
   type UserModerationActionVo,
 } from '@/api/moderationApi';
-import { getApiBaseUrl } from '@/config/env';
 import { CONSOLE_PERMISSIONS } from '@/constants/permissions';
 import { useDocumentTitle } from '@/hooks/useDocumentTitle';
 import { usePermission } from '@/hooks/usePermission';
 import { log } from '@/utils/logger';
+import {
+  ACTION_LOG_ACTION_TYPE_OPTIONS,
+  ACTION_LOG_STATUS_OPTIONS,
+  ACTION_OPTIONS,
+  MANUAL_ACTION_OPTIONS,
+  MANUAL_ACTION_TYPE,
+  REASON_TYPE_OPTIONS,
+  REVIEW_STATUS_OPTIONS,
+  TARGET_NAVIGATION_STATUS_OPTIONS,
+  TARGET_TYPE_OPTIONS,
+  buildActionLogManualActionReason,
+  buildActionSourceTargetDisplayInput,
+  buildActionSourceTargetNavigationInput,
+  buildManualModerationStatusSnapshot,
+  buildQueueManualActionReason,
+  buildQueueTargetDisplayInput,
+  buildQueueTargetNavigationInput,
+  canOpenChatTarget,
+  getActionTypeText,
+  getManualActionTypeText,
+  getReasonTypeLabel,
+  renderActionType,
+  renderModerationTarget,
+  renderReportStatus,
+  resolveMissingTargetMessage,
+  resolveOpenTarget,
+  toOptionalString,
+  toPositiveLongString,
+  type ActionLogActiveFilter,
+  type ActionLogPreset,
+  type ManualActionPreset,
+  type ManualActionTypeValue,
+  type ManualActiveModerationStatus,
+  type ManualModerationStatusSnapshot,
+  type ModerationTargetNavigationStateInput,
+  type QueuePreset,
+} from './moderationPageHelpers';
 import './index.css';
 import '../adminFeature.css';
-
-const REVIEW_STATUS_OPTIONS = [
-  { label: '全部状态', value: -1 },
-  { label: '待审核', value: 0 },
-  { label: '已通过', value: 1 },
-  { label: '已驳回', value: 2 },
-];
-
-const TARGET_TYPE_OPTIONS = [
-  { label: '帖子', value: 'Post' },
-  { label: '评论', value: 'Comment' },
-  { label: '轻回应', value: 'PostQuickReply' },
-  { label: '聊天消息', value: 'ChatMessage' },
-  { label: '商品', value: 'Product' },
-];
-
-const REASON_TYPE_OPTIONS = [
-  { label: '垃圾内容', value: 'Spam' },
-  { label: '辱骂骚扰', value: 'Abuse' },
-  { label: '色情低俗', value: 'Pornography' },
-  { label: '违法违规', value: 'Illegal' },
-  { label: '欺诈误导', value: 'Fraud' },
-  { label: '其他', value: 'Other' },
-];
-
-const TARGET_NAVIGATION_STATUS_OPTIONS = [
-  { label: '可回看', value: 'Ready' },
-  { label: '已降级', value: 'Fallback' },
-  { label: '已失效', value: 'Unavailable' },
-  { label: '暂不支持', value: 'Unsupported' },
-];
-
-const ACTION_OPTIONS = [
-  { label: '不处罚', value: 0 },
-  { label: '禁言', value: 1 },
-  { label: '封禁', value: 2 },
-];
-
-const MANUAL_ACTION_TYPE = {
-  mute: 1,
-  ban: 2,
-  unmute: 3,
-  unban: 4,
-} as const;
-
-type ManualActionTypeValue = typeof MANUAL_ACTION_TYPE[keyof typeof MANUAL_ACTION_TYPE];
-
-const MANUAL_ACTION_OPTIONS = [
-  { label: '禁言', value: MANUAL_ACTION_TYPE.mute },
-  { label: '封禁', value: MANUAL_ACTION_TYPE.ban },
-  { label: '解除禁言', value: MANUAL_ACTION_TYPE.unmute },
-  { label: '解除封禁', value: MANUAL_ACTION_TYPE.unban },
-];
-
-const ACTION_LOG_ACTION_TYPE_OPTIONS = [
-  { label: '禁言', value: 'Mute' },
-  { label: '封禁', value: 'Ban' },
-  { label: '解除禁言', value: 'Unmute' },
-  { label: '解除封禁', value: 'Unban' },
-];
-
-const ACTION_LOG_STATUS_OPTIONS = [
-  { label: '生效中', value: 'active' },
-  { label: '已结束', value: 'inactive' },
-];
-
-const TARGET_TYPE_LABEL_MAP: Record<string, string> = {
-  Post: '帖子',
-  Comment: '评论',
-  PostQuickReply: '轻回应',
-  ChatMessage: '聊天消息',
-  Product: '商品',
-  Unknown: '未知目标',
-};
-
-const REASON_TYPE_LABEL_MAP: Record<string, string> = {
-  Spam: '垃圾内容',
-  Abuse: '辱骂骚扰',
-  Pornography: '色情低俗',
-  Illegal: '违法违规',
-  Fraud: '欺诈误导',
-  Other: '其他',
-};
-
-function getTargetTypeLabel(value: string | null | undefined): string {
-  return value ? (TARGET_TYPE_LABEL_MAP[value] ?? value) : '未知目标';
-}
-
-function getReasonTypeLabel(value: string | null | undefined): string {
-  return value ? (REASON_TYPE_LABEL_MAP[value] ?? value) : '未分类';
-}
-
-function toOptionalString(value: unknown): string | undefined {
-  return typeof value === 'string' && value.trim().length > 0 ? value.trim() : undefined;
-}
-
-function toPositiveLongString(value: string): string | undefined {
-  const trimmed = value.trim();
-  return /^[1-9]\d*$/.test(trimmed) ? trimmed : undefined;
-}
-
-const renderReportStatus = (value: string) => {
-  switch (value) {
-    case 'Approved':
-      return <Tag color="success">已通过</Tag>;
-    case 'Rejected':
-      return <Tag color="error">已驳回</Tag>;
-    default:
-      return <Tag color="processing">待审核</Tag>;
-  }
-};
-
-const renderActionType = (value: string) => {
-  switch (value) {
-    case 'Mute':
-      return <Tag color="orange">禁言</Tag>;
-    case 'Ban':
-      return <Tag color="red">封禁</Tag>;
-    case 'Unmute':
-      return <Tag color="green">解除禁言</Tag>;
-    case 'Unban':
-      return <Tag color="cyan">解除封禁</Tag>;
-    default:
-      return <Tag>无动作</Tag>;
-  }
-};
-
-function getActionTypeText(value: string | null | undefined): string {
-  switch (value) {
-    case 'Mute':
-      return '禁言';
-    case 'Ban':
-      return '封禁';
-    case 'Unmute':
-      return '解除禁言';
-    case 'Unban':
-      return '解除封禁';
-    default:
-      return '治理动作';
-  }
-}
-
-function getManualActionTypeText(value: ManualActionTypeValue | null | undefined): string {
-  switch (value) {
-    case MANUAL_ACTION_TYPE.mute:
-      return '禁言';
-    case MANUAL_ACTION_TYPE.ban:
-      return '封禁';
-    case MANUAL_ACTION_TYPE.unmute:
-      return '解除禁言';
-    case MANUAL_ACTION_TYPE.unban:
-      return '解除封禁';
-    default:
-      return '治理动作';
-  }
-}
-
-function buildDesktopChatTargetUrl(channelId: number, messageId: number): string {
-  const url = new URL('/desktop', getApiBaseUrl());
-  url.searchParams.set('app', 'chat');
-  url.searchParams.set('channelId', String(channelId));
-  url.searchParams.set('messageId', String(messageId));
-  return url.toString();
-}
-
-function buildPublicForumTargetUrl(postId: number, commentId?: number | null): string {
-  const url = new URL(`/forum/post/${postId}`, getApiBaseUrl());
-  if (commentId && commentId > 0) {
-    url.searchParams.set('commentId', String(commentId));
-  }
-
-  return url.toString();
-}
-
-function buildPublicShopTargetUrl(productId: number): string {
-  return new URL(`/shop/product/${productId}`, getApiBaseUrl()).toString();
-}
-
-function canOpenChatTarget(targetType: string | null | undefined, channelId: number | null | undefined, messageId: number | null | undefined): boolean {
-  return targetType === 'ChatMessage'
-    && !!channelId
-    && channelId > 0
-    && !!messageId
-    && messageId > 0;
-}
-
-interface ModerationTargetNavigationInput {
-  targetType: string | null | undefined;
-  targetContentId?: number | null;
-  targetPostId?: number | null;
-  targetCommentId?: number | null;
-  targetChannelId?: number | null;
-  targetMessageId?: number | null;
-}
-
-interface ModerationOpenTarget {
-  label: string;
-  url: string;
-}
-
-interface ModerationTargetNavigationStateInput extends ModerationTargetNavigationInput {
-  navigationStatus?: string | null;
-  navigationMessage?: string | null;
-}
-
-interface ModerationTargetDisplayInput extends ModerationTargetNavigationStateInput {
-  snapshotTitle?: string | null;
-  snapshotSummary?: string | null;
-  snapshotIsPersisted?: boolean;
-  targetUserId?: number | null;
-  targetUserName?: string | null;
-  showTargetUser?: boolean;
-}
-
-type ActionLogActiveFilter = 'active' | 'inactive';
-
-interface ActionLogPreset {
-  targetUserId?: string;
-  sourceReportId?: string;
-  actionType?: string;
-  isActive?: ActionLogActiveFilter;
-  keyword?: string;
-  hint: string;
-}
-
-interface QueuePreset {
-  status?: number;
-  targetType?: string;
-  reasonType?: string;
-  navigationStatus?: string;
-  keyword?: string;
-  hint: string;
-}
-
-interface ManualActionPreset {
-  targetUserId?: string;
-  sourceReportId?: string;
-  actionType?: ManualActionTypeValue;
-  durationHours?: number | null;
-  reason?: string;
-  hint: string;
-}
-
-interface ManualActiveModerationStatus {
-  actionId: number;
-  actionType: 'Mute' | 'Ban';
-  reason: string;
-  sourceReportId?: number | null;
-  endTime?: string | null;
-}
-
-interface ManualModerationStatusSnapshot {
-  targetUserId: string;
-  muteAction: ManualActiveModerationStatus | null;
-  banAction: ManualActiveModerationStatus | null;
-}
-
-function resolveNavigationStatusLabel(status: string | null | undefined): { color: string; label: string } {
-  switch (status) {
-    case 'Fallback':
-      return { color: 'warning', label: '已降级' };
-    case 'Unavailable':
-      return { color: 'default', label: '已失效' };
-    case 'Unsupported':
-      return { color: 'default', label: '暂不支持' };
-    default:
-      return { color: 'success', label: '可回看' };
-  }
-}
-
-function renderTargetNavigationState(status: string | null | undefined, messageText: string | null | undefined) {
-  const statusMeta = resolveNavigationStatusLabel(status);
-
-  return (
-    <div className="moderation-target__section">
-      <div className="moderation-target__section-label">当前状态</div>
-      <div className="moderation-target__state">
-        <Tag color={statusMeta.color}>{statusMeta.label}</Tag>
-        {messageText ? <div className="moderation-target__state-message">{messageText}</div> : null}
-      </div>
-    </div>
-  );
-}
-
-function renderSnapshotSection(input: ModerationTargetDisplayInput) {
-  const hasSnapshotTitle = !!input.snapshotTitle?.trim();
-  const hasSnapshotSummary = !!input.snapshotSummary?.trim();
-
-  if (!hasSnapshotTitle && !hasSnapshotSummary && !input.snapshotIsPersisted) {
-    return null;
-  }
-
-  return (
-    <div className="moderation-target__section">
-      <div className="moderation-target__section-label">
-        {input.snapshotIsPersisted ? '创建时快照' : '目标摘要（旧数据兼容）'}
-      </div>
-      {hasSnapshotTitle ? <div className="moderation-target__snapshot-title">{input.snapshotTitle}</div> : null}
-      {hasSnapshotSummary ? <div className="moderation-target__snapshot-summary">{input.snapshotSummary}</div> : null}
-      {!hasSnapshotTitle && !hasSnapshotSummary ? <div className="moderation-target__empty">未保留文本摘要</div> : null}
-    </div>
-  );
-}
-
-function renderModerationTarget(input: ModerationTargetDisplayInput) {
-  return (
-    <div className="moderation-target">
-      <div className="moderation-target__identity">{getTargetTypeLabel(input.targetType)} #{input.targetContentId ?? '-'}</div>
-      {input.targetType === 'Comment' && input.targetPostId ? (
-        <div className="moderation-target__meta">
-          帖子 #{input.targetPostId} · 评论 #{input.targetCommentId ?? input.targetContentId}
-        </div>
-      ) : null}
-      {input.targetType === 'PostQuickReply' && input.targetPostId ? (
-        <div className="moderation-target__meta">所属帖子 #{input.targetPostId}</div>
-      ) : null}
-      {input.targetType === 'ChatMessage' && input.targetChannelId ? (
-        <div className="moderation-target__meta">
-          频道 #{input.targetChannelId} · 消息 #{input.targetMessageId ?? input.targetContentId}
-        </div>
-      ) : null}
-      {renderSnapshotSection(input)}
-      {renderTargetNavigationState(input.navigationStatus, input.navigationMessage)}
-      {input.showTargetUser ? (
-        <div className="moderation-target__user">{input.targetUserName || `用户 ${input.targetUserId}`}</div>
-      ) : null}
-    </div>
-  );
-}
-
-function buildQueueTargetDisplayInput(record: ContentReportQueueItemVo): ModerationTargetDisplayInput {
-  return {
-    targetType: record.voTargetType,
-    targetContentId: record.voTargetContentId,
-    targetPostId: record.voTargetPostId,
-    targetCommentId: record.voTargetCommentId,
-    targetChannelId: record.voTargetChannelId,
-    targetMessageId: record.voTargetMessageId,
-    navigationStatus: record.voTargetNavigationStatus,
-    navigationMessage: record.voTargetNavigationMessage,
-    snapshotTitle: record.voTargetSnapshotTitle,
-    snapshotSummary: record.voTargetSnapshotSummary,
-    snapshotIsPersisted: record.voTargetSnapshotIsPersisted,
-    targetUserId: record.voTargetUserId,
-    targetUserName: record.voTargetUserName,
-  };
-}
-
-function buildQueueTargetNavigationInput(record: ContentReportQueueItemVo): ModerationTargetNavigationStateInput {
-  return {
-    targetType: record.voTargetType,
-    targetContentId: record.voTargetContentId,
-    targetPostId: record.voTargetPostId,
-    targetCommentId: record.voTargetCommentId,
-    targetChannelId: record.voTargetChannelId,
-    targetMessageId: record.voTargetMessageId,
-    navigationStatus: record.voTargetNavigationStatus,
-    navigationMessage: record.voTargetNavigationMessage,
-  };
-}
-
-function buildActionSourceTargetDisplayInput(record: UserModerationActionVo): ModerationTargetDisplayInput {
-  return {
-    targetType: record.voSourceReportTargetType,
-    targetContentId: record.voSourceReportTargetContentId ?? null,
-    targetPostId: record.voSourceReportTargetPostId,
-    targetCommentId: record.voSourceReportTargetCommentId,
-    targetChannelId: record.voSourceReportTargetChannelId,
-    targetMessageId: record.voSourceReportTargetMessageId,
-    navigationStatus: record.voSourceReportTargetNavigationStatus,
-    navigationMessage: record.voSourceReportTargetNavigationMessage,
-    snapshotTitle: record.voSourceReportTargetSnapshotTitle,
-    snapshotSummary: record.voSourceReportTargetSnapshotSummary,
-    snapshotIsPersisted: record.voSourceReportTargetSnapshotIsPersisted,
-  };
-}
-
-function buildActionSourceTargetNavigationInput(record: UserModerationActionVo): ModerationTargetNavigationStateInput {
-  return {
-    targetType: record.voSourceReportTargetType,
-    targetContentId: record.voSourceReportTargetContentId,
-    targetPostId: record.voSourceReportTargetPostId,
-    targetCommentId: record.voSourceReportTargetCommentId,
-    targetChannelId: record.voSourceReportTargetChannelId,
-    targetMessageId: record.voSourceReportTargetMessageId,
-    navigationStatus: record.voSourceReportTargetNavigationStatus,
-    navigationMessage: record.voSourceReportTargetNavigationMessage,
-  };
-}
-
-function buildQueueManualActionReason(record: ContentReportQueueItemVo): string {
-  const reasonLabel = getReasonTypeLabel(record.voReasonType);
-  const reasonDetail = record.voReasonDetail?.trim();
-  return reasonDetail
-    ? `来源举报单 #${record.voReportId}：${reasonLabel}，${reasonDetail}`
-    : `来源举报单 #${record.voReportId}：${reasonLabel}`;
-}
-
-function buildActionLogManualActionReason(record: UserModerationActionVo, actionType?: ManualActionTypeValue): string {
-  if (actionType === MANUAL_ACTION_TYPE.unmute || actionType === MANUAL_ACTION_TYPE.unban) {
-    return `参考动作单 #${record.voActionId}，人工复核后${getManualActionTypeText(actionType)}`;
-  }
-
-  return `参考动作单 #${record.voActionId}，继续对用户 #${record.voTargetUserId} 执行人工治理`;
-}
-
-function pickLatestActiveModerationAction(
-  actions: UserModerationActionVo[],
-  actionType: 'Mute' | 'Ban'
-): ManualActiveModerationStatus | null {
-  const matchedAction = actions
-    .filter(action => action.voIsActive && action.voActionType === actionType)
-    .sort((left, right) => new Date(right.voStartTime).getTime() - new Date(left.voStartTime).getTime())[0];
-
-  if (!matchedAction) {
-    return null;
-  }
-
-  return {
-    actionId: matchedAction.voActionId,
-    actionType,
-    reason: matchedAction.voReason,
-    sourceReportId: matchedAction.voSourceReportId,
-    endTime: matchedAction.voEndTime,
-  };
-}
-
-function buildManualModerationStatusSnapshot(
-  targetUserId: string,
-  actions: UserModerationActionVo[]
-): ManualModerationStatusSnapshot {
-  return {
-    targetUserId,
-    muteAction: pickLatestActiveModerationAction(actions, 'Mute'),
-    banAction: pickLatestActiveModerationAction(actions, 'Ban'),
-  };
-}
-
-function resolveOpenTarget(input: ModerationTargetNavigationStateInput): ModerationOpenTarget | null {
-  const navigationStatus = input.navigationStatus ?? 'Ready';
-  if (navigationStatus === 'Unavailable' || navigationStatus === 'Unsupported') {
-    return null;
-  }
-
-  const targetType = input.targetType ?? null;
-  if (canOpenChatTarget(targetType, input.targetChannelId, input.targetMessageId)) {
-    return {
-      label: '打开聊天定位',
-      url: buildDesktopChatTargetUrl(Number(input.targetChannelId), Number(input.targetMessageId)),
-    };
-  }
-
-  if (targetType === 'Post' && input.targetPostId && input.targetPostId > 0) {
-    return {
-      label: '打开帖子详情',
-      url: buildPublicForumTargetUrl(Number(input.targetPostId)),
-    };
-  }
-
-  if (targetType === 'Comment' && input.targetPostId && input.targetPostId > 0) {
-    const targetCommentId = input.targetCommentId ?? input.targetContentId;
-    const isFallback = navigationStatus === 'Fallback';
-    return {
-      label: isFallback ? '打开所属帖子' : '打开评论定位',
-      url: buildPublicForumTargetUrl(Number(input.targetPostId), isFallback ? undefined : targetCommentId),
-    };
-  }
-
-  if (targetType === 'PostQuickReply' && input.targetPostId && input.targetPostId > 0) {
-    return {
-      label: '打开所属帖子',
-      url: buildPublicForumTargetUrl(Number(input.targetPostId)),
-    };
-  }
-
-  if (targetType === 'Product' && input.targetContentId && input.targetContentId > 0) {
-    return {
-      label: '打开商品详情',
-      url: buildPublicShopTargetUrl(Number(input.targetContentId)),
-    };
-  }
-
-  return null;
-}
-
-function resolveMissingTargetMessage(targetType: string | null | undefined, navigationMessage?: string | null): string {
-  if (navigationMessage && navigationMessage.trim().length > 0) {
-    return navigationMessage;
-  }
-
-  switch (targetType) {
-    case 'ChatMessage':
-      return '当前举报项缺少聊天定位信息';
-    case 'Post':
-    case 'Comment':
-    case 'PostQuickReply':
-      return '当前举报项缺少论坛定位信息';
-    case 'Product':
-      return '当前举报项缺少商品定位信息';
-    default:
-      return '当前举报项暂不支持直接回看';
-  }
-}
 
 export const ModerationPage = () => {
   useDocumentTitle('内容治理');
